+++ packages/astro/src/core/add/babel.ts
@@ -3,12 +3,12 @@ import parser from '@babel/parser';
 import traverse from '@babel/traverse';
 import * as t from '@babel/types';
 
-// @ts-ignore @babel/traverse isn't ESM and needs this trick
+// @ts-expect-error @babel/traverse isn't ESM and needs this trick
 export const visit = traverse.default as typeof traverse;
 export { t };
 
 export async function generate(ast: t.File) {
-	// @ts-ignore @babel/generator isn't ESM and needs this trick
+	// @ts-expect-error @babel/generator isn't ESM and needs this trick
 	const astToText = generator.default as typeof generator;
 	const { code } = astToText(ast);
 	return code;