+++ packages/astro/src/core/errors/errors-data.ts
@@ -75,7 +75,7 @@ export const AstroErrorData = defineErrors({
 	 * @description
 	 * The `Astro.clientAddress` property is only available when [Server-side rendering](https://docs.astro.build/en/guides/server-side-rendering/) is enabled.
 	 *
-	 * To get the user's IP address in static mode, different APIs such as [Ipify](https://www.ipify.org/) can be used in a [Client-side script](https://docs.astro.build/en/core-concepts/astro-components/#client-side-scripts) or it may be possible to get the user's IP using a serverless function hosted on your hosting provider.
+	 * To get the user's IP address in static mode, different APIs such as [Ipify](https://www.ipify.org/) can be used in a [Client-side script](https://docs.astro.build/en/guides/client-side-scripts/) or it may be possible to get the user's IP using a serverless function hosted on your hosting provider.
 	 */
 	StaticClientAddressNotAvailable: {
 		title: '`Astro.clientAddress` is not available in static mode.',
@@ -453,6 +453,84 @@ See https://docs.astro.build/en/guides/server-side-rendering/ for more informati
 		message: (paramName: string) =>
 			`[paginate()] page number param \`${paramName}\` not found in your filepath.`,
 		hint: 'Rename your file to `[page].astro` or `[...page].astro`.',
+	},
+	/**
+	 * @docs
+	 * @see
+	 * - [Assets (Experimental)](https://docs.astro.build/en/guides/assets/)
+	 * - [Image component](https://docs.astro.build/en/guides/assets/#image--astroassets)
+	 * - [Image component#alt](https://docs.astro.build/en/guides/assets/#alt-required)
+	 * @description
+	 * The `alt` property allows you to provide descriptive alt text to users of screen readers and other assistive technologies. In order to ensure your images are accessible, the `Image` component requires that an `alt` be specified.
+	 *
+	 * If the image is merely decorative (i.e. doesn’t contribute to the understanding of the page), set `alt=""` so that screen readers know to ignore the image.
+	 */
+	ImageMissingAlt: {
+		title: 'Missing alt property',
+		code: 3022,
+		message: 'The alt property is required.',
+		hint: "The `alt` property is important for the purpose of accessibility, without it users using screen readers or other assistive technologies won't be able to understand what your image is supposed to represent. See https://developer.mozilla.org/en-US/docs/Web/HTML/Element/img#attr-alt for more information.",
+	},
+	/**
+	 * @docs
+	 * @see
+	 * - [Image Service API](https://docs.astro.build/en/reference/image-service-reference/)
+	 * @description
+	 * There was an error while loading the configured image service. This can be caused by various factors, such as your image service not properly exporting a compatible object in its default export, or an incorrect path.
+	 *
+	 * If you believe that your service is properly configured and this error is wrong, please [open an issue](https://astro.build/issues/).
+	 */
+	InvalidImageService: {
+		title: 'Error while loading image service',
+		code: 3023,
+		message:
+			'There was an error loading the configured image service. Please see the stack trace for more information.',
+	},
+	/**
+	 * @docs
+	 * @message
+	 * Missing width and height attributes for `IMAGE_URL`. When using remote images, both dimensions are always required in order to avoid cumulative layout shift (CLS).
+	 * @see
+	 * - [Assets (Experimental)](https://docs.astro.build/en/guides/assets/)
+	 * - [Image component#width-and-height](https://docs.astro.build/en/guides/assets/#width-and-height)
+	 * @description
+	 * For remote images, `width` and `height` cannot be inferred from the original file. As such, in order to avoid CLS, those two properties are always required.
+	 *
+	 * If your image is inside your `src` folder, you probably meant to import it instead. See [the Imports guide for more information](https://docs.astro.build/en/guides/imports/#other-assets).
+	 */
+	MissingImageDimension: {
+		title: 'Missing image dimensions',
+		code: 3024,
+		message: (missingDimension: 'width' | 'height' | 'both', imageURL: string) =>
+			`Missing ${
+				missingDimension === 'both'
+					? 'width and height attributes'
+					: `${missingDimension} attribute`
+			} for ${imageURL}. When using remote images, both dimensions are always required in order to avoid CLS.`,
+		hint: 'If your image is inside your `src` folder, you probably meant to import it instead. See [the Imports guide for more information](https://docs.astro.build/en/guides/imports/#other-assets).',
+	},
+	/**
+	 * @docs
+	 * @description
+	 * The built-in image services do not currently support optimizing all image formats.
+	 *
+	 * For unsupported formats such as SVGs and GIFs, you may be able to use an `img` tag directly:
+	 * ```astro
+	 * ---
+	 * import rocket from '../assets/images/rocket.svg'
+	 * ---
+	 *
+	 * <img src={rocket.src} width={rocket.width} height={rocket.height} alt="A rocketship in space." />
+	 * ```
+	 */
+	UnsupportedImageFormat: {
+		title: 'Unsupported image format',
+		code: 3025,
+		message: (format: string, imagePath: string, supportedFormats: readonly string[]) =>
+			`Received unsupported format \`${format}\` from \`${imagePath}\`. Currently only ${supportedFormats.join(
+				', '
+			)} are supported for optimization.`,
+		hint: "If you do not need optimization, using an `img` tag directly instead of the `Image` component might be what you're looking for.",
 	},
 	// No headings here, that way Vite errors are merged with Astro ones in the docs, which makes more sense to users.
 	// Vite Errors - 4xxx