+++ packages/astro/test/test-utils.js
@@ -12,6 +12,7 @@ import { createSettings } from '../dist/core/config/index.js';
 import dev from '../dist/core/dev/index.js';
 import { nodeLogDestination } from '../dist/core/logger/node.js';
 import preview from '../dist/core/preview/index.js';
+import { check } from '../dist/cli/check/index.js';
 
 // polyfill WebAPIs to globalThis for Node v12, Node v14, and Node v16
 polyfill(globalThis, {
@@ -24,6 +25,8 @@ polyfill(globalThis, {
  * @typedef {import('../src/@types/astro').AstroConfig} AstroConfig
  * @typedef {import('../src/core/preview/index').PreviewServer} PreviewServer
  * @typedef {import('../src/core/app/index').App} App
+ * @typedef {import('../src/cli/check/index').AstroChecker} AstroChecker
+ * @typedef {import('../src/cli/check/index').CheckPayload} CheckPayload
  *
  *
  * @typedef {Object} Fixture
@@ -39,6 +42,24 @@ polyfill(globalThis, {
  * @property {() => Promise<void>} clean
  * @property {() => Promise<App>} loadTestAdapterApp
  * @property {() => Promise<void>} onNextChange
+ * @property {(opts: CheckPayload) => Promise<AstroChecker>} check
+ *
+ * This function returns an instance of the Check
+ *
+ *
+ * When used in a test suite:
+ * ```js
+ * let fixture = await loadFixture({
+ *   root: './fixtures/astro-check-watch/',
+ * });
+ * ```
+ * `opts` will override the options passed to the `AstroChecker`
+ *
+ * ```js
+ * let { check, stop, watch } = fixture.check({
+ *   flags: { watch: true },
+ * });
+ * ```
  */
 
 /** @type {import('../src/core/logger/core').LogOptions} */
@@ -148,6 +169,9 @@ export async function loadFixture(inlineConfig) {
 			return build(settings, { logging, telemetry, ...opts });
 		},
 		sync: (opts) => sync(settings, { logging, fs, ...opts }),
+		check: async (opts) => {
+			return await check(settings, { logging, ...opts });
+		},
 		startDevServer: async (opts = {}) => {
 			process.env.NODE_ENV = 'development';
 			devServer = await dev(settings, { logging, telemetry, ...opts });
@@ -160,7 +184,11 @@ export async function loadFixture(inlineConfig) {
 		fetch: (url, init) => fetch(resolveUrl(url), init),
 		preview: async (opts = {}) => {
 			process.env.NODE_ENV = 'production';
-			const previewServer = await preview(settings, { logging, telemetry, ...opts });
+			const previewServer = await preview(settings, {
+				logging,
+				telemetry,
+				...opts,
+			});
 			config.server.host = parseAddressToHost(previewServer.host); // update host
 			config.server.port = previewServer.port; // update port
 			return previewServer;
@@ -177,7 +205,11 @@ export async function loadFixture(inlineConfig) {
 				cwd: fileURLToPath(config.outDir),
 			}),
 		clean: async () => {
-			await fs.promises.rm(config.outDir, { maxRetries: 10, recursive: true, force: true });
+			await fs.promises.rm(config.outDir, {
+				maxRetries: 10,
+				recursive: true,
+				force: true,
+			});
 		},
 		loadTestAdapterApp: async (streaming) => {
 			const url = new URL(`./server/entry.mjs?id=${fixtureId}`, config.outDir);
@@ -250,7 +282,9 @@ const cliPath = fileURLToPath(new URL('../astro.js', import.meta.url));
 
 /** Returns a process running the Astro CLI. */
 export function cli(/** @type {string[]} */ ...args) {
-	const spawned = execa('node', [cliPath, ...args], { env: { ASTRO_TELEMETRY_DISABLED: true } });
+	const spawned = execa('node', [cliPath, ...args], {
+		env: { ASTRO_TELEMETRY_DISABLED: true },
+	});
 
 	spawned.stdout.setEncoding('utf8');