+++ packages/integrations/preact/src/client-dev.ts
@@ -1,4 +1,3 @@
-// @ts-ignore
 import 'preact/debug';
 import clientFn from './client.js';